// save-mate/src/app/core/enums/contact-category.enum.ts
export enum ContactCategory {
    generalInquiry = 'Általános kérdés',
    productServiceInquiry = 'Érdeklődés szolgáltatásról / termékről',
    technicalIssue = 'Technikai probléma',
    bugReport = 'Hibabejelentés',
    partnershipInquiry = 'Együttműködés / Partneri megkeresés',
    pressMediaInquiry = 'Sajtó / Média megkeresés',
    feedbackComplaint = 'Visszajelzés / Panasz',
    other = 'Egyéb'
}